[#2434] Make package names in kafka-common to adhere the common pattern.
