+++ clients/adapter-kafka/src/main/java/org/eclipse/hono/adapter/client/telemetry/kafka/KafkaBasedEventSender.java
@@ -17,10 +17,10 @@ import java.util.Map;
 import java.util.Objects;
 
 import org.eclipse.hono.adapter.client.telemetry.EventSender;
+import org.eclipse.hono.client.kafka.HonoTopic;
+import org.eclipse.hono.client.kafka.KafkaProducerConfigProperties;
+import org.eclipse.hono.client.kafka.KafkaProducerFactory;
 import org.eclipse.hono.config.ProtocolAdapterProperties;
-import org.eclipse.hono.kafka.client.HonoTopic;
-import org.eclipse.hono.kafka.client.KafkaProducerConfigProperties;
-import org.eclipse.hono.kafka.client.KafkaProducerFactory;
 import org.eclipse.hono.util.EventConstants;
 import org.eclipse.hono.util.QoS;
 import org.eclipse.hono.util.RegistrationAssertion;
