+++ clients/kafka-common/src/main/java/org/eclipse/hono/client/kafka/consumer/AbstractAtLeastOnceKafkaConsumer.java
@@ -11,7 +11,7 @@
  * SPDX-License-Identifier: EPL-2.0
  */
 
-package org.eclipse.hono.kafka.client.consumer;
+package org.eclipse.hono.client.kafka.consumer;
 
 import java.time.Duration;
 import java.util.List;
