+++ clients/kafka-common/src/main/java/org/eclipse/hono/client/kafka/consumer/KafkaConsumerCommitException.java
@@ -11,19 +11,19 @@
  * SPDX-License-Identifier: EPL-2.0
  */
 
-package org.eclipse.hono.kafka.client.consumer;
+package org.eclipse.hono.client.kafka.consumer;
 
 /**
- * An exception indicating that consuming data from Kafka failed.
+ * An exception indicating that a Kafka consumer failed to commit offsets.
  */
-public class KafkaConsumerException extends RuntimeException {
+public class KafkaConsumerCommitException extends KafkaConsumerException {
 
     /**
      * Creates a new exception for a root cause.
      *
      * @param cause The root cause.
      */
-    public KafkaConsumerException(final Throwable cause) {
+    public KafkaConsumerCommitException(final Throwable cause) {
         super(cause);
     }
