+++ clients/kafka-common/src/main/java/org/eclipse/hono/client/kafka/tracing/KafkaHeadersInjectAdapter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 Contributors to the Eclipse Foundation
+ * Copyright (c) 2020, 2021 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -11,10 +11,8 @@
  * SPDX-License-Identifier: EPL-2.0
  */
 
-package org.eclipse.hono.kafka.client.tracing;
+package org.eclipse.hono.client.kafka.tracing;
 
-import java.util.AbstractMap;
-import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map.Entry;
@@ -24,10 +22,10 @@ import io.opentracing.propagation.TextMap;
 import io.vertx.kafka.client.producer.KafkaHeader;
 
 /**
- * An adapter for extracting properties from a list of {@link KafkaHeader} objects.
+ * An adapter for injecting properties into a list of {@link KafkaHeader} objects.
  *
  */
-public final class KafkaHeadersExtractAdapter implements TextMap {
+public final class KafkaHeadersInjectAdapter implements TextMap {
 
     private final List<KafkaHeader> headers;
 
@@ -37,33 +35,17 @@ public final class KafkaHeadersExtractAdapter implements TextMap {
      * @param headers The list of {@link KafkaHeader} objects.
      * @throws NullPointerException if headers is {@code null}.
      */
-    public KafkaHeadersExtractAdapter(final List<KafkaHeader> headers) {
+    public KafkaHeadersInjectAdapter(final List<KafkaHeader> headers) {
         this.headers = Objects.requireNonNull(headers);
     }
 
     @Override
     public Iterator<Entry<String, String>> iterator() {
-        if (headers.isEmpty()) {
-            return Collections.emptyIterator();
-        }
-        final Iterator<KafkaHeader> headersIterator = headers.iterator();
-        return new Iterator<>() {
-
-            @Override
-            public boolean hasNext() {
-                return headersIterator.hasNext();
-            }
-
-            @Override
-            public Entry<String, String> next() {
-                final KafkaHeader header = headersIterator.next();
-                return new AbstractMap.SimpleEntry<>(header.key(), header.value().toString());
-            }
-        };
+        throw new UnsupportedOperationException();
     }
 
     @Override
     public void put(final String key, final String value) {
-        throw new UnsupportedOperationException();
+        headers.add(KafkaHeader.header(key, value));
     }
 }
