+++ clients/kafka-common/src/test/java/org/eclipse/hono/client/kafka/HonoTopicTest.java
@@ -11,7 +11,7 @@
  * SPDX-License-Identifier: EPL-2.0
  */
 
-package org.eclipse.hono.kafka.client;
+package org.eclipse.hono.client.kafka;
 
 import static org.assertj.core.api.Assertions.assertThat;
