+++ clients/kafka-common/src/test/java/org/eclipse/hono/client/kafka/RecordUtilTest.java
@@ -11,7 +11,7 @@
  * SPDX-License-Identifier: EPL-2.0
  *******************************************************************************/
 
-package org.eclipse.hono.kafka.util;
+package org.eclipse.hono.client.kafka;
 
 import static org.assertj.core.api.Assertions.assertThat;
